Added loading option to view all game logs table. Also implemented an error if you don't have any game logs. Added the pop up for the delete button

+++ web/src/pages/viewGameLog.js
@@ -76,17 +76,23 @@ class ViewGameLog extends BindingClass {
         const deleteButton = document.getElementById('deleteGameLogButton');
         const messageContainer = document.getElementById('messageContainer');
         const deleteMessage = document.getElementById('message');
-
-        this.client.deleteGameLog(gameLog.gameId);
-
-        deleteButton.style.display = 'none';
-        messageContainer.style.display = 'block';
-
-        deleteMessage.textContent = 'GameLog has been deleted';
-
-        setTimeout(() => {
-            window.location.href = `/deleteGameLog.html`;
-          }, 3000); //3000 milli = 3 sec
+    
+        const confirmed = confirm("Are you sure you want to delete the game log?");
+    
+        if (confirmed) {
+            this.client.deleteGameLog(gameLog.gameId);
+    
+            deleteButton.style.display = 'none';
+            messageContainer.style.display = 'block';
+    
+            deleteMessage.textContent = 'GameLog has been deleted';
+    
+            setTimeout(() => {
+                window.location.href = `/deleteGameLog.html`;
+            }, 3000); //3000 milliseconds = 3 seconds
+        } else {
+            console.log("Deletion canceled");
+        }
     }
 }
